% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-interface.R
\name{add_track_block}
\alias{add_track_block}
\title{Add Track With Different Types of Features}
\usage{
add_track_block(tnt, data, label, height = 40, color.background = "white",
  color.feature = "black")
}
\arguments{
\item{tnt}{}

\item{data}{}

\item{label}{}

\item{height}{}

\item{color.background}{}

\item{color.feature}{}
}
\examples{
if (interactive()) {
    mydata1 <- data.frame (
        start = c(42, 69, 233),
        end = c(54, 99, 250)
    )
    mydata2 <- data.frame(
        start = c(23, 66, 300),
        end = c(38, 74, 318)
    )
    tnt_board(from = 14, to = 114, min = -100, max = 500) \%>\%
        add_track_block(mydata1, label = "My Track 1") \%>\%
        add_track_block(mydata2, label = "My Track 2", color.feature = "green") \%>\%
        TnT()
}
}

