% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JScascade.R
\name{asJS}
\alias{asJS}
\title{Parse \code{\link{JScascade}} Class as Literal Javascript Code}
\usage{
asJS(x)
}
\arguments{
\item{x}{}
}
\value{
It will return a character string of the class "JS_EVAL".
}
\description{
Parse \code{\link{JScascade}} Class as Literal Javascript Code
}
