% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JScascade.R
\name{JScascade}
\alias{JScascade}
\alias{jc}
\title{Construct a Javascript Function Cascade}
\usage{
JScascade(...)

jc(...)
}
\arguments{
\item{...}{}
}
\value{
The function returns a JScascade S3 object which can be further utilized
    by \code{asJS()} and \code{JScascade()}/\code{jc()} itself.
}
\description{
\code{JScascade} constructs a list which can represent the defination of a
TnT browser/tree as a functional cascade, it can then be parsed as javascript
code by \code{\link{asJS}}.
}
\examples{
axisTrack <- jc(
    tnt.board.track = NULL,
    height = 20,
    color = "white",
    display = jc(tnt.board.track.feature.axis = NULL)
)
blockTrack <- jc(
    tnt.board.track = NULL,
    height = 30,
    color = "yellow",
    data = jc(
        tnt.board.track.data.sync = NULL,
        retriever = JS("function() {return [{start : 200, end : 350}]}")
    ),
    display = jc(
        tnt.board.track.feature.block = NULL,
        color = "blue",
        index = JS("function (d) {return d.start}")
    )
)
tntdef <- jc(
    tnt.board = NULL,
    from = 0,
    to = 500,
    min = 50,
    max = 1000,
    width = 500,
    add_track = axisTrack,
    add_track = blockTrack
)
tntdef
asJS(tntdef)
TnT(asJS(tntdef))
}

