% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JScascade.R
\name{quo}
\alias{equo}
\alias{quo}
\title{Add Quotation Mark to a String}
\usage{
quo(string)

equo(x)
}
\arguments{
\item{string}{}

\item{x}{It will be subsitituted and deparsed as a string. Then \code{equo} will
call \code{quo} on it returning a string with additional quotation marks.}
}
\value{

}
\description{
Utility to facilitate the construction of \code{\link{JScascade}}.
}
\examples{
identical(
    quo("white"),
    equo(white)
)
q <- quo("hell('world')")
cat(q)
if (requireNamespace("jsonlite"))
    jsonlite::toJSON(q)
}

