% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tntBoard.R
\docType{methods}
\name{range,TnTBoard-method}
\alias{range,TnTBoard-method}
\title{Range of TnTBoard}
\usage{
\S4method{range}{TnTBoard}(x, ..., with.revmap = FALSE,
  ignore.strand = FALSE, na.rm = FALSE)
}
\arguments{
\item{x}{TnTBoard.}

\item{..., with.revmap, ignore.strand, na.rm}{Passed to \code{\link[GenomicRanges]{range,GenomicRanges-method}}.}
}
\value{
GRanges.
}
\description{
Get combined range of all tracks in a TnTBoard, used internally.
}
