% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{strandlabel}
\alias{strandlabel}
\title{Display Labels with Strand}
\usage{
strandlabel(labels, strands)
}
\arguments{
\item{labels}{Character vector, names of each feature.}

\item{strands}{Factor or character vector with the same length of \code{labels}, can
be "+", "-" or "*".}
}
\value{
A character vector that combines the labels with strand information.
}
\description{
A simple util function that used internally to generate display labels of GeneTrack and
TxTrack.
}
\examples{
strandlabel(c("gene1", "gene2", "gene3"), c("+", "-", "*"))
}
