% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tntTracks.R
\docType{methods}
\name{tooltip}
\alias{tooltip}
\alias{tooltip<-}
\alias{tooltip,TrackData-method}
\alias{tooltip,TnTTrack-method}
\alias{tooltip<-,TrackData,data.frame-method}
\alias{tooltip<-,TnTTrack,data.frame-method}
\title{Access Track Tooltips}
\usage{
tooltip(x)

tooltip(x) <- value

\S4method{tooltip}{TrackData}(x)

\S4method{tooltip}{TnTTrack}(x)

\S4method{tooltip}{TrackData,data.frame}(x) <- value

\S4method{tooltip}{TnTTrack,data.frame}(x) <- value
}
\arguments{
\item{x}{A TnTTrack object.}

\item{value}{A data frame to replace, its row number should equal to length of data.}
}
\value{
\code{tooltip} returns a data frame.
}
\description{
Access Track Tooltips
}
\examples{
gr <- GRanges("chr12", IRanges(c(6, 69), c(42, 135)), Name = c("my range 1", "my range 2"))
track <- BlockTrack(gr)
tooltip(track)
tooltip(track)$Width <- width(gr)
tooltip(track)
}
