% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tntTracks.R
\docType{methods}
\name{trackdata}
\alias{trackdata}
\alias{trackData}
\alias{trackData<-}
\alias{$,TnTTrack-method}
\alias{$<-,TnTTrack-method}
\title{Access Track Data}
\usage{
trackData(x)

trackData(x) <- value

\S4method{$}{TnTTrack}(x, name)

\S4method{$}{TnTTrack}(x, name) <- value
}
\arguments{
\item{x}{A TnTTrack object.}

\item{value}{Replaced value.}

\item{name}{Passed to the inner method for track data.}
}
\description{
Access and modify the track data. \code{x$name} and \code{x$name <- value} are
just shortcuts for \code{trackData(x)$name} and \code{trackData(x)$name <- value},
respectively.
}
\examples{
track <- BlockTrack(GRanges("chr1", IRanges(6, 54)))
trackData(track) # track data of block track is an object that inherits GRanges.
ctrack <- merge(track, track)
trackData(ctrack) # track data of composite track is a list of tracks
}
