% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tntBoard.R
\name{tracklist}
\alias{tracklist}
\alias{tracklist<-}
\title{Track List in TnTBoard}
\usage{
tracklist(tntboard)

tracklist(tntboard) <- value
}
\arguments{
\item{tntboard}{A TnTBoard or TnTGenome object}

\item{value}{A list of tracks}
}
\description{
The tracks of a TnTBoard are stored as a list which can be accessed or modified
with these functions.
}
\examples{
bt <- BlockTrack(GRanges("chr21", IRanges(100, 1200)))
li.tracks <- list(bt, bt)
board <- TnTBoard(li.tracks)
tracklist(board)
if (interactive()) show(board)
tracklist(board) <- list(bt)
if (interactive()) show(board)
}
