% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JScascade.R
\name{JScascade}
\alias{JScascade}
\alias{jc}
\title{Construct a Javascript Function Cascade}
\usage{
JScascade(...)

jc(...)
}
\arguments{
\item{...}{}
}
\value{
The function returns a JScascade S3 object which can be further utilized
    by \code{asJS()} and \code{JScascade()}/\code{jc()} itself.
}
\description{
\code{JScascade} constructs a list which can represent the defination of a
TnT browser/tree as a functional cascade, it can then be parsed as javascript
code by \code{\link{asJS}}.
}
\examples{
axisTrack <- jc(
    tnt.board.track = "",
    height = 20,
    color = quo('white'),
    display = jc(tnt.board.track.feature.axis = "")
)
blockTrack <- jc(
    tnt.board.track = "",
    height = 30,
    color = quo("#FFCFDD"),
    data = jc(
        tnt.board.track.data.sync = "",
        retriever = "function() {return [{start : 200, end : 350}]}"
    ),
    display = jc(
        tnt.board.track.feature.block = "",
        color = quo("blue"),
        index = "function (d) {return d.start}"
    )
)
tntdef <- jc(
    tnt.board.genome = "",
    from = 0,
    to = 500,
    min = 0,
    max = 1000,
    add_track = axisTrack,
    add_track = blockTrack
)
tntdef
asJS(tntdef)
}

