% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tntTracks-compositeTrack.R
\name{range-TnTTrack}
\alias{range-TnTTrack}
\alias{range,RangeBasedTrack-method}
\alias{range,CompositeTrack-method}
\title{Range of Tracks}
\usage{
\S4method{range}{RangeBasedTrack}(x, ..., with.revmap = FALSE, ignore.strand = FALSE, na.rm = FALSE)

\S4method{range}{CompositeTrack}(x, ..., with.revmap = FALSE, ignore.strand = FALSE, na.rm = FALSE)
}
\arguments{
\item{x}{A TnTTrack object.}

\item{..., with.revmap, ignore.strand, na.rm}{Passed to \code{\link[GenomicRanges]{range,GenomicRanges-method}}.}
}
\value{
Returns a GRanges object.
}
\description{
Range of Tracks
}
