% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{crunch}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{biovizBase}{\code{\link[biovizBase]{crunch}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

